+++ Classes/Domain/Validator/RegisterValidator.php
@@ -21,21 +21,21 @@ class RegisterValidator extends AbstractValidator
      * @var \Wacon\Feuserregistration\Domain\Service\Validation\RegisterValidationService
      */
     protected $registerValidationService;
-    
+
     protected function isValid($value): void
     {
-        $this->registerValidationService = GeneralUtility::makeInstance(\Wacon\Feuserregistration\Domain\Service\Validation\RegisterValidationService::class);                
+        $this->registerValidationService = GeneralUtility::makeInstance(\Wacon\Feuserregistration\Domain\Service\Validation\RegisterValidationService::class);
         $className = get_class($value);
 
-        if ($className != \Wacon\Feuserregistration\Domain\Model\User::class) {
+        if (($value instanceof \Wacon\Feuserregistration\Domain\Model\User) !== true) {
             $errorString = 'The user validator can only handle object of class ' . \Wacon\Feuserregistration\Domain\Model\User::class . ', '
-                . $className . ' given instead.';
+                . get_class($value). ' given instead.';
             $this->addError($errorString, time());
         }
 
         if (!$this->registerValidationService->isValid($value)) {
             $propertiesWithError = $this->registerValidationService->getPropertiesWithError();
-            
+
             foreach($propertiesWithError as $propertyWithError) {
                 $this->addErrorForProperty($propertyWithError['name'], $propertyWithError['errorString'], $propertyWithError['errorCode']);
             }