+++ Tests/Domain/Model/ExtendedUser.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+/*
+ * This file is part of the TYPO3 extension feuserregistration.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ *  (c) 2023 Kevin Chileong Lee, info@wacon.de, WACON Internet GmbH
+ */
+
+namespace Wacon\Feuserregistration\Tests\Domain\Model;
+
+class ExtendedUser extends \Wacon\Feuserregistration\Domain\Model\User
+{}