+++ Tests/Domain/Model/NotExtendedUser.php
@@ -1,0 +1,18 @@
+<?php
+declare(strict_types=1);
+
+/*
+ * This file is part of the TYPO3 extension feuserregistration.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ *  (c) 2023 Kevin Chileong Lee, info@wacon.de, WACON Internet GmbH
+ */
+
+namespace Wacon\Feuserregistration\Tests\Domain\Model;
+
+use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
+
+class NotExtendedUser extends AbstractEntity
+{}